rename package: attributeInfo => attributes
